feat: Modal 컴포넌트에 Escape 키로 닫기 기능 추가 및 useModal 훅에서 관련 코드 제거
